pyproject.toml
[tool.poetry]
name = "python-notes"
version = "0.1.0"
description = "Python Notes"
authors = [
    "Darren Weber <dazza-codes@gitlab.com>",
]
license = "Apache-2.0"

readme = "README.md"  # Markdown files are supported
repository = "https://gitlab.com/dazza-codes/python-notes.git"
homepage = "https://gitlab.com/dazza-codes/python-notes"

keywords = [
    "Development Status :: 2 - Pre-Alpha",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Topic :: Utilities",
]

include = ["notes/**/*"]
exclude = ["tests/**/*"]

packages = [
    {include = "notes"}
]

# [tool.poetry.scripts]
# hello = "notes.__main__:main"

[tool.poetry.dependencies]
python = "^3.6"
dataclasses = {version = "*", python = "^3.6.0"}

aiobotocore = {extras = ["awscli", "boto3"], version = "^0.12.0"}
aioredis = "^1.3.1"
click = "^7.0"
databases = {extras = ["postgresql", "mysql", "sqlite"], version = "*"}
tinydb = "^3.15.2"

[tool.poetry.dev-dependencies]
autopep8 = "*"
black = "^19.10b0"
codecov = "*"
doc8 = "*"
flake8 = "*"
flake8-type-annotations = "*"
ipdb = "*"
invoke-release = "*"
m2r = "*"
moto = {extras = ["server"], version = "^1.3.14" }
mypy = "*"
pylint = "*"
pytest = "*"
pytest-asyncio = "*"
pytest-benchmark = "*"
pytest-cov = "*"
pytest-datadir = "*"
pytest-mock = "*"
pytest-profiling = "*"
pytest-runner = "*"
pytest-xdist = "*"
readme-renderer = { version = "*", extras = ["md"] }
setuptools = "*"
sphinx = "*"
sphinx-autoapi = "*"
sphinx_rtd_theme = "*"
twine = "*"

[tool.black]
line-length = 96

[build-system]
requires = ["poetry>=1.0"]
build-backend = "poetry.masonry.api"
